UBUNTU: SAUCE: Use ticket locks for Xen 3.0.2+
The SUSE patchset did limit the ticket lock code to be used only
on Xen hypervisors version 3.2 or higher because HYPERVISOR_poll()
is supposed to have issues before.
However we have been using that call before and the chance to be
running on EC2 on such an old hypervisor is getting even less.
Not using ticket locks and the hypervisor assisting could increase
host cpu usage and unfairness in lock contention situations.

BugLink: http://bugs.launchpad.net/bugs/929941

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Herton Krzesinski <herton.krzesinski@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>